separated logging to it's own package
